git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@9680 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ admin/includes/classes/import.php
@@ -39,7 +39,7 @@ class xtcImport {
   **
   *****************************************************************************/
 
-    function xtcImport($filename) {
+    function __construct($filename) {
         $this->seperator = CSV_SEPERATOR;
         $this->TextSign = CSV_TEXTSIGN;
         if (trim(CSV_TEXTSIGN) == '') {
@@ -53,7 +53,7 @@ class xtcImport {
         }
         $this->filename = $filename;
         $this->ImportDir = DIR_FS_CATALOG.'import/';
-        $this->catDepth = isset(CSV_CAT_DEPTH) ? CSV_CAT_DEPTH : 4;
+        $this->catDepth = defined(CSV_CAT_DEPTH) ? CSV_CAT_DEPTH : 4;
         $this->languages = $this->get_lang();
         $this->counter = array ('prod_new' => 0,
                                 'cat_new' => 0,
@@ -69,6 +69,10 @@ class xtcImport {
         $this->CatDefault = CSV_CATEGORY_DEFAULT;
         $this->FileSheme = array ();
         $this->Groups = xtc_get_customers_statuses();
+        //BOC new var to use in for loops, noRiddle
+        $this->count_groups = count($this->Groups);
+        $this->sizeof_languages = sizeof($this->languages);
+        //EOC new var to use in for loops, noRiddle
     }
 
     /*****************************************************************************
@@ -107,11 +111,11 @@ class xtcImport {
           'p_shipping' => '' // product shipping_time
         );
         // group prices
-        for ($i = 0; $i < count($this->Groups) - 1; $i ++) {
+        for ($i = 0; $i < $this->count_groups - 1; $i ++) {
             $file_layout = array_merge($file_layout, array ('p_priceNoTax.'.$this->Groups[$i +1]['id'] => ''));
         }
         // group permissions
-        for ($i = 0; $i < count($this->Groups); $i ++) {
+        for ($i = 0; $i < $this->count_groups; $i ++) {
             $file_layout = array_merge($file_layout, array ('p_groupAcc.'.$this->Groups[$i]['id'] => ''));
         }
         // product images
@@ -119,7 +123,7 @@ class xtcImport {
             $file_layout = array_merge($file_layout, array ('p_image.'.$i => ''));
         }
         // add lang fields
-        for ($i = 0; $i < sizeof($this->languages); $i ++) {
+        for ($i = 0; $i < $this->sizeof_languages; $i ++) {
             $file_layout = array_merge($file_layout, array ('p_name.'.$this->languages[$i]['code'] => '',
                                                             'p_desc.'.$this->languages[$i]['code'] => '',
                                                             'p_shortdesc.'.$this->languages[$i]['code'] => '',
@@ -432,11 +436,10 @@ class xtcImport {
             $products_id = $prod_data['products_id'];
         }
 
-        // Insert group prices.
-
-        for ($i = 0; $i < count($this->Groups) - 1; $i ++) {
+        // Insert group prices Qantity:Price::Quantity:Price
+        for ($i = 0; $i < $this->count_groups - 1; $i ++) {
             // seperate string ::
-            if (isset ($dataArray['p_priceNoTax.'.$this->Groups[$i +1]['id']])) {
+            if (isset ($dataArray['p_priceNoTax.'.$this->Groups[$i +1]['id']]) && $dataArray['p_priceNoTax.'.$this->Groups[$i +1]['id']] != '') {
 
                 $truncate_query = "DELETE FROM ".TABLE_PERSONAL_OFFERS_BY.$this->Groups[$i +1]['id']."
                                          WHERE products_id = '".$products_id."'";
@@ -444,7 +447,7 @@ class xtcImport {
                 xtc_db_query($truncate_query);
                 $prices = $dataArray['p_priceNoTax.'.$this->Groups[$i +1]['id']];
                 $prices = explode('::', $prices);
-                for ($ii = 0; $ii < count($prices); $ii ++) {
+                for ($ii = 0, $cp = count($prices); $ii < $cp; $ii ++) {
                     $values = explode(':', $prices[$ii]);
 
                     $group_array = array (
@@ -459,7 +462,7 @@ class xtcImport {
         }
 
         // Insert group permissions.
-        for ($i = 0; $i < count($this->Groups); $i ++) {
+        for ($i = 0; $i < $this->count_groups; $i ++) {
             if (isset ($dataArray['p_groupAcc.'.$this->Groups[$i]['id']])) {
                 $insert_array = array ('group_permission_'.$this->Groups[$i]['id'] => $dataArray['p_groupAcc.'.$this->Groups[$i]['id']]);
                 xtc_db_perform(TABLE_PRODUCTS, $insert_array, 'update', 'products_id = \''.$products_id.'\'');
@@ -486,7 +489,7 @@ class xtcImport {
 
         if ($touchCat) $this->insertCategory($dataArray, $mode, $products_id);
 
-        for ($i_insert = 0; $i_insert < sizeof($this->languages); $i_insert ++) {
+        for ($i_insert = 0; $i_insert < $this->sizeof_languages; $i_insert ++) {
             $prod_desc_array = array (
                 'products_id' => $products_id,
                 'language_id' => $this->languages[$i_insert]['id']
@@ -568,7 +571,7 @@ class xtcImport {
             $catTree = '';
             $parTree = '';
             $curr_ID = 0;
-            for ($i = 0; $i < count($cat); $i ++) {
+            for ($i = 0, $cc = count($cat); $i < $cc; $i ++) {
                 $catTree .= '[\''.xtc_db_input($cat[$i]).'\']';
                 $code = '$ID=$this->CatTree'.$catTree.'[\'ID\'];';
                 //debug
@@ -593,10 +596,10 @@ class xtcImport {
                     $cat_query = xtc_db_query("SELECT c.categories_id
                                                  FROM ".TABLE_CATEGORIES." c
                                                  JOIN ".TABLE_CATEGORIES_DESCRIPTION." cd
-                                                      ON cd.categories_id=c.categories_id
-                                                         AND cd.language_id='".$this->languages[0]['id']."'
-                                                         AND cd.categories_name='".addslashes($cat[$i])."'
-                                                WHERE c.parent_id='".$parent."'");
+                                                   ON cd.categories_id = c.categories_id
+                                                WHERE cd.categories_name = '".xtc_db_input($cat[$i])."'
+                                                  AND cd.language_id = '".$this->languages[0]['id']."'
+                                                  AND c.parent_id = '".$parent."'");
                     if (!xtc_db_num_rows($cat_query)) { // insert categorie
                         $categorie_data = array (
                             'parent_id' => $parent,
@@ -619,7 +622,7 @@ class xtcImport {
                         if ($this->debug) echo '<pre>FIRST $CAT_ID: '.$cat_id.'</pre><hr />';
 
                         $parent = $cat_id;
-                        for ($i_insert = 0; $i_insert < sizeof($this->languages); $i_insert ++) {
+                        for ($i_insert = 0; $i_insert < $this->sizeof_languages; $i_insert ++) {
                             $categorie_data = array (
                                 'language_id' => $this->languages[$i_insert]['id'],
                                 'categories_id' => $cat_id,
@@ -776,8 +779,8 @@ class xtcExport {
     **
     *****************************************************************************/
 
-    function xtcExport($filename) {
-        $this->catDepth = isset(CSV_CAT_DEPTH) ? CSV_CAT_DEPTH : 4;
+    function __construct($filename) {
+        $this->catDepth = defined(CSV_CAT_DEPTH) ? CSV_CAT_DEPTH : 4;
         $this->languages = $this->get_lang();
         $this->filename = $filename;
         $this->CAT = array ();
@@ -798,6 +801,10 @@ class xtcExport {
             $this->seperator = "\t";
 
         $this->Groups = xtc_get_customers_statuses();
+        //BOC new var to use in for loops, noRiddle
+        $this->count_groups = count($this->Groups);
+        $this->sizeof_languages = sizeof($this->languages);
+        //EOC new var to use in for loops, noRiddle
     }
 
     /*****************************************************************************
@@ -876,11 +883,11 @@ class xtcExport {
             $line .= $this->encode($heading);
         }
 
-        for ($i=1; $i<count($this->Groups); $i++) {
+        for ($i=1; $i < $this->count_groups; $i++) {
             $line .= $this->encode('p_priceNoTax.'.$this->Groups[$i]['id']);
         }
         if (GROUP_CHECK == 'true') {
-            for ($i=0; $i<count($this->Groups); $i++) {
+            for ($i=0; $i < $this->count_groups; $i++) {
                 $line .= $this->encode('p_groupAcc.'.$this->Groups[$i]['id']);
             }
         }
@@ -907,7 +914,7 @@ class xtcExport {
         $line .= $this->encode('p_image');
 
         // add lang fields
-        for ($i = 0; $i < sizeof($this->languages); $i ++) {
+        for ($i = 0; $i < $this->sizeof_languages; $i ++) {
             $line .= $this->encode('p_name.'.$this->languages[$i]['code']);
             $line .= $this->encode('p_desc.'.$this->languages[$i]['code']);
             $line .= $this->encode('p_shortdesc.'.$this->languages[$i]['code']);
@@ -944,7 +951,7 @@ class xtcExport {
             $line .= $this->encode($export_data['products_price']);
 
             // group prices  Qantity:Price::Quantity:Price
-            for ($i=1; $i<count($this->Groups); $i++) {
+            for ($i=1; $i < $this->count_groups; $i++) {
                 $price_query = "SELECT *
                                   FROM ".TABLE_PERSONAL_OFFERS_BY.$this->Groups[$i]['id']."
                                  WHERE products_id = '".$export_data['products_id']."'
@@ -969,7 +976,7 @@ class xtcExport {
 
             // group permissions
             if (GROUP_CHECK == 'true') {
-                for ($i=0; $i<count($this->Groups); $i++) {
+                for ($i=0; $i < $this->count_groups; $i++) {
                     $line .= $this->encode($export_data['group_permission_'.$this->Groups[$i]['id']]);
                 }
             }
@@ -1009,7 +1016,7 @@ class xtcExport {
 
             $line .= $this->encode($export_data['products_image']);
 
-            for ($i = 0; $i < sizeof($this->languages); $i ++) {
+            for ($i = 0; $i < $this->sizeof_languages; $i ++) {
                 $lang_query = xtc_db_query("SELECT *
                                               FROM ".TABLE_PRODUCTS_DESCRIPTION."
                                              WHERE language_id = '".$this->languages[$i]['id']."'
